refactor(harness): rename cluster tests, add hopr-transport::testing module
Rename `stress_cluster-*` → `cluster_throughput-*` across all hopr-lib test
entries and the example binary to reflect what the harness actually does: a
real-QUIC multi-node cluster throughput exercise over a mock Blokli chain (not
"all stubbed"). Fix the stale `--nodes` flag references in the example doc
(parser only accepts `--hops`).

Generalize the common bench/test utilities into a proper `testing` feature on
`hopr-transport` (`src/testing/stubs.rs` + `src/testing/harness.rs`), eliminating
the `#[path = "../tests/protocol/common/mod.rs"]` hack from both benches. Both
`pipeline_e2e_bench` and `protocol_throughput_emulated_bench` now import from
`hopr_transport::testing`; the old `common/mod.rs` becomes a thin re-export shim
so existing test call sites continue compiling unchanged.
