+++ hopr/hopr-lib/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "hopr-lib"
 description = "HOPR library implementing the HOPR protocol"
-version = "4.0.2-rc.7"
+version = "4.0.2-rc.8"
 edition.workspace = true
 license.workspace = true
 readme = "README.md"
@@ -29,7 +29,7 @@ name = "chain_operations-size3"
 required-features = ["testing"]
 
 [[test]]
-name = "stress_cluster-size3"
+name = "cluster_throughput-size3"
 required-features = ["testing"]
 
 [[test]]
@@ -53,15 +53,15 @@ name = "transport_tickets"
 required-features = ["testing"]
 
 [[test]]
-name = "stress_cluster-size5"
+name = "cluster_throughput-size5"
 required-features = ["testing"]
 
 [[test]]
-name = "stress_cluster-matrix"
+name = "cluster_throughput-matrix"
 required-features = ["testing"]
 
 [[example]]
-name = "stress_cluster"
+name = "cluster_throughput"
 required-features = ["testing"]
