+++ transport/hopr/benches/protocol_throughput_emulated_bench.rs
@@ -1,10 +1,7 @@
-#[path = "../tests/protocol/common/mod.rs"]
-mod common;
-
 use std::{str::FromStr, sync::Arc};
 
 use bytes::Bytes;
-use common::{CHAIN_DATA, PEERS, PEERS_CHAIN, random_packets_of_count, resolve_mock_path};
+use hopr_transport::testing::harness::{CHAIN_DATA, PEERS, PEERS_CHAIN, random_packets_of_count, resolve_mock_path};
 use criterion::{BenchmarkId, Criterion, Throughput, criterion_group, criterion_main};
 use futures::{SinkExt, StreamExt};
 use hopr_api::{
